// src/types/types.ts
export interface UseCardFormProps {
    card: {id:string,title:string,description:string,priority:string}
    onEditCard: (cardId: string, columnId: string, newCard: { title: string; description: string; priority: string }) => void;
    columnId: string
}

export interface UseCardFormResult {
    isEditing: boolean;
    title: string;
    description: string;
    priority: string;
    setTitle: (value: string) => void;
    setDescription: (value: string) => void;
    setPriority: (value: string) => void;
    handleEditClick: () => void;
    handleSaveClick: () => void;
    handleCancelClick: () => void;
}